Add a `delete_state` function for branching
Skip the test for branching to give way for the implementation of
the `delete_state`. The function `delete_state` will be used to
incrementally delete all the states so that the initial and
final states are the only left. The labels between the states before
the state the selected state and after will be a regex that has the
same effect when that state is present.