Make it work when there are three states
The start state in the original is not included in the reverse because
of the final state having no nodes to point to. So we will assume that
the keys of the original and the reverse are equal. If there are
states in the set of next states that are not in the original, it is
wrong by default and should throw an error.