Implement a full version of deletion of a state
This will now delete a state and connect all possible paths from the
previous states to the next states of the state in consideration.
This will also detects the looping so that it will not be considered
when detecting the possible paths.

It goes like this
     0  1,-v 2
-> a ---> b ---> c
If b is deleted, the resulting regex should be 01*2 because it will
accept a 0, then many number of one (which can be none) and then 2.