Implement the conversion of automata to regex
Delete all the states that are not initial or final. After that we
have an automata which is composed of only one state (of which can loop
to itself) or two state automata.

In case of two state automata, we can have the final regex for free
by adding an state before the start state and state after the final
state and delete the start and final states. The pre-start state will
have the regex of the whole automata. We assume here that the
pre-starting state will have only one resulting state. It would be an
error if it is not otherwise.