Merge the two cases of the final form of automata
Remove the code that is supposed to be on the `delete_state` from the
`auto2regex`. Having the pre and post states are useful even for the
case of length 1. It will also give the final regex form of the
automata.

Rearrange the order of adding state and deletion so that the case of
one state automata will pass the test. Deleting the state before adding
the post-final state will not give the correct answer if that state is
also the final state. The pre-start state will not have the combined
regex of the starting/final state.